Refactored vimgmt function naming, updated gitignore and readme
While refactoring everything recently to use the name "vimgmt", I forgot
that vim doesn't allow plugin functions to start with a lowercase
letter, so I had to go through and update the functions to begin with
"Vimgmt".

Gitignore was updated to exclude the resulting vimball from manually
building.

Readme was updated with the new function names, as well as a small FAQ
section.