Finished GL support in vimgmt.sh, updated readme
Updated the vimgmt script with finished support for gitlab issues.
Accessing project level information like issues, pull requests, etc
requires the project id (not the repo name), so to get the project id,
there has to be an initial curl request.

In order to match most of the vimscript code for parsing github issues,
I also had to add a lot more jq filtering of the returned response json
from gitlab. Namely, "description" -> "body", "iid" -> "number", and
transforming the labels list of strings to a list of objects in the same
format as those from github.

The reamde was also updated with a comment on why token encryption is
necessary.