Added support for issue navigation and commenting
Selecting an issue (by hitting enter) now creates a new temporary buffer
with all issue info, as well as available comments. To return to the
main view, you can use :VimgmtBack.

Commenting is now supported, by running :VimgmtComment. At the moment,
only single line comments are supported. Eventually it would be
preferable to create an entirely new buffer for the user to write out a
comment, and then exiting the buffer would submit the body of text as
the comment body.

Issue details and commenting are handled with an entirely new script,
vimgmt_issue.sh.

The scripts have been updated to include support for a custom title,
body, and labels as well.

There is also now a VimgmtClose function, which has not been completed
yet. It should function as a way to close issues and pull requests.