Improved main view navigation, reduced header size
The main view now has a much smaller size, as body text is not shown for
each issue. This more closely mimics the behavior seen on GitHub/GitLab,
and allows the user to quickly navigate between issues.

Selecting an issue now opens a new buffer to the right, if there is room
for one.

The header text was replaced with a smaller banner, to reduce the amount
of clutter in the main view.