Improved comment workflow, refactored function names
Comment workflow was improved to now use :VimgmtComment to open a
comment buffer, and :VimgmtPost to post the contents of the comment
buffer as a new comment. This allows for multiline comments, code
blocks, etc to be posted in a comment.

Function names were refactored to use "Create" instead of "Make" as the
primary verb, which is mostly just a personal preference.