Add :VimgmtJump to home page navigation
:VimgmtJump takes one argument and allows the user to jump between
issues on the home page. '1' moves the cursor to the next issue, and
'-1' moves the cursor to the previous issue.

Added buffer-only mappings for 'gn' and 'gp' to jump to the
next/previous issue respectively. Also added 'gi' mapping as a shortcut
for :VimgmtBack, similar to the behavior on GitHub's website.