Simplify response caching
The directory declaration used for calling scripts was refactored to be
a global variable that can be used for caching responses from the API.
This simplifies things a bit as it means all responses/tokens/etc reside
in the same directory, and allows the bash scripts to encrypt and write the
responses directly from the server rather than having vim do it as an
additional step.

Each command in the bash script was updated to store the final response
as a new variable, encrypt and write the contents of the variable to the
cache, and echo back the response to vim. The plugin can then decrypt
and read the files as needed.