Refactor shell scripts to vimscript only (partial)
This is a partially complete refactoring of functionality achieved with
shell scripting to be achieved with pure vimscript instead. The primary
intent of this is to reduce the amount of dependencies that a user may
require when installing vim-repoman.

Effectively, this will remove the need for a bash interpreter, and it
will remove the jq dependency, since all result formatting/filtering
will be done in vimscript.