Add repo list navigation when not in a repo
When the user is not in a repository, the plugin will now pull a list of
repositories for the authenticated user, with the ability to jump
between repos and enter/exit them similarly to the issue list view.

If the user is in a repo, the plugin functions the same.

Also added a little bit of (temporary) syntax highlighting to color
separation bars between items. This should be updated to pull the
comment color from the user's current syntax theme.