Support updating labels from list, simplify jump guide
Labels can now be updated from the main issue list, which makes sense as
a feature since labels are visible from the list and don't require
anything beyond an issue number to be updated.

The jump guide was simplified to use change in cursor position to
determine if the next jump should cycle to beginning/end of the list.

Also fixed vint issue of string comparison during decoration syntax
highlighting.