Add ability to edit comments, update comment UI
Comments can now be edited via Ctrl-e when hovered over a comment. This
opens a buffer containing the current comment body, and executing
RepoManPost from this buffer calls the API for updating the comment.

Comment barriers in the UI were updated to be a bit more noticeable,
although the downside of this is that the syntax highlighting call
breaks the comment character highlighting. May need to revisit this
again if it becomes annoying.