Fix nav back in 1-buf view, fix sanitize text
The "navigate back" functionality was updated to check for a currently
opened issue before deciding which buffer to switch to. This fixes the
bug that would occur in single buffer issue views, which would navigate
back to the repo list if available (rather than the issue list view for
the repo).

The text sanitization utility was updated to only conditionally escape
double quotes, since that is only required when passing data in curl
requests.