Restructure buffer class, update request return type
Request class now calls the curl request and returns a json decoded
element, rather than relying on the GitHub API class to do both of those
steps for each call.

Buffer class was restructured to distinguish between functions that
require an updated plugin state, and functions that are purely utility
based buffer functions.

Documentation was updated for a lot of methods to indicate args and
return type.