Refactor buffer mgmt (#18)
* Move buffer functions to separate class

Buffer operations are being moved to their own class to reduce the
amount of clutter in the primary repoman.vim script (which should only
handle logic related to user input).

Constants were also moved to their own file.

* Restructure buffer class, update request return type

Request class now calls the curl request and returns a json decoded
element, rather than relying on the GitHub API class to do both of those
steps for each call.

Buffer class was restructured to distinguish between functions that
require an updated plugin state, and functions that are purely utility
based buffer functions.

Documentation was updated for a lot of methods to indicate args and
return type.

* Fix select repo from repo list

Selecting a repo in the repo list to inspect required an update to the
arguments for CreateIssuesListBuffer.

* Update script desc fields, add missing scriptencoding

Script encoding was missing from the constants script.

Description fields were added/updated for every script.