Use pr diff url as pr indicator
Using a true/false for indicating if the user is in a PR wasn't very
helpful beyond just ui components. The indicator has been updated to
store the pr diff url instead, which can be used for fetching the diff
and displaying it during review.

Note that since vimscript doesn't evaluate non-empty strings as true
(why?) I have to append a '1' to the beginning of the url. This should
obviously be removed when fetching the diff, but -- while janky --
eliminates a lot of extra work for evaluating the string as
empty/non-empty every time.