Add support for adding review comments
Using the same RepoManComment -> RepoManPost workflow seen in issues,
comments can now be added for particular changes in a code review. This
updates the UI with a new block below the particular line and inserts
the contents of the comment (multline supported), shifting the remaining
contents of the diff down depending on the number of lines in the
comment.

The next step should be to add the comments to a dictionary for later
use when submitting the code review to GitHub.