Add support for multiline comments/suggestions
Note that multiline comments and single line comments require completely
different parameters for telling GitHub where the comment should appear.
For single-line comments, the 'position' value is required, which is
determined by the number of lines below the first @@ header for each
file in the diff. Multiline comments require the 'start_[line|side]' and
'[line|side]' parameters, which use the actual file line number.

The two can't be mixed, as the GitHub API disallows setting the start
and end line number to the same line to create a single line comment.
(Which imo would be a lot simpler, but ¯\_(ツ)_/¯)