Add support for PR reviews (#20)
* Add user functions for beginning code review

Adds :RepoManReview [new|comment|approve|request_changes]
Adds :RepoManSave

Logic for calling RepoManReview has been implemented to ensure a valid
state.

* Use pr diff url as pr indicator

Using a true/false for indicating if the user is in a PR wasn't very
helpful beyond just ui components. The indicator has been updated to
store the pr diff url instead, which can be used for fetching the diff
and displaying it during review.

Note that since vimscript doesn't evaluate non-empty strings as true
(why?) I have to append a '1' to the beginning of the url. This should
obviously be removed when fetching the diff, but -- while janky --
eliminates a lot of extra work for evaluating the string as
empty/non-empty every time.

* Add ability to reply to review comments

* Add ability to view pr diff, update curl return type

PR diffs can now be opened in a buffer, which is done using
":RepoManReview new".

The curl request script now returns the raw system response, since the
new diff request can't be json decoded.

* Add review buffer lookup table

This introduces a new buffer variable for looking up the file and
(relative to the diff) line number, which can be used for adding
comments.

* Add support for adding review comments

Using the same RepoManComment -> RepoManPost workflow seen in issues,
comments can now be added for particular changes in a code review. This
updates the UI with a new block below the particular line and inserts
the contents of the comment (multline supported), shifting the remaining
contents of the diff down depending on the number of lines in the
comment.

The next step should be to add the comments to a dictionary for later
use when submitting the code review to GitHub.

* Add ability to remove and edit review buffer comments

* Add ability to delete review comments

Also adds a user facing method - RepoManDelete - which can be used for
regular comments as well.

* Add ability to write code suggestions in review

Adds RepoManSuggest range method

* Add support for multiline comments/suggestions

Note that multiline comments and single line comments require completely
different parameters for telling GitHub where the comment should appear.
For single-line comments, the 'position' value is required, which is
determined by the number of lines below the first @@ header for each
file in the diff. Multiline comments require the 'start_[line|side]' and
'[line|side]' parameters, which use the actual file line number.

The two can't be mixed, as the GitHub API disallows setting the start
and end line number to the same line to create a single line comment.
(Which imo would be a lot simpler, but ¯\_(ツ)_/¯)

* Improve review reply formatting in details view

Review replies were not visually distinct enough from regular review
comments, so a new "UI" was designed in its stead. The new design has
more of a bi-level tree design, with top level comments appearing
closest to the edge of the buffer, and all replies appearing indented
beneath the review comment. Subsequent replies do not increase the
indent further.

* Add support for review comment/request_changes

* Update documentation

* Remove firstline use from reply function