Add basic ability to edit issues
Only includes editing title and body, but it is technically "editing"
nonetheless. This uses the same RepoManEdit command seen with comments,
but when executed from anywhere on an issue details view besides over a
comment, allows editing of the issue details instead.

In the future, this should be expanded upon to differentiate between
issues and pull requests, and allow editing of other details, such as
assignees, milestones, etc.

See #21