Refactor project to new name (vim-hubbub)
After an unnecessary amount of deliberation for a project that only I
and maybe one other person uses, I've decided to rename everything from
"repoman" to "hubbub". The reasons for the rename I'll outline below.

The plugin does "manage" repos, but the idea of "managing" a repo makes
it sound kinda dull. It more provides a makeshift GitHub UI inside of
Vim, which can be useful for anyone who doesn't like using their browser
for simple things like commenting on an issue. So the former "repoman"
name for the plugin wasn't doing any favors for the project by implying
that it's only for people who want to "manage" repos in GitHub.

The old name also just seemed a bit too edgelord for me, and had
confusing capitalization in documentation (Repoman vs RepoMan).