Remove plugin name from header
The plugin name being included in the header introduced unnecessary
clutter to the view, and made the main "repos" list view confusing. Now
the header only states the current repo name and page number, if the
plugin is opened from within a repo, or a "Repositories (page #)" header
if opened outside of a repo.